// src/types/credit.ts
export interface ICredit {
  group_name_ru: string;
  inf_id: string | number;
  inf_koe: string | number;
  inf_max_size: number;
  inf_odolg: string | number;
  inf_oproc: string | number;
  inf_proc_formula: number;
  inf_time: number;
  kredit_type: string;
  platName: string;
  usl_name: string;
  val_key: string;
}